Fix order export if order detail `taxID` is null
When exporting an order that contains an item whose `taxID` is `null` and `tax_rate` is zero, the export fails due to an exception, because the DB adapter tries to find a tax rate using `null` as query parameter. Orders created using the Shopware checkout for some reason set `0` instead `null` for the `taxID`, if no tax is mapped, and hence never trigger this error. Based on the database schema of `s_order_details` it is fine though to set `taxID` to `null`.

This PR adds an extra check to `MainOrdersDbAdapter::calculateTaxSum()` to prevent using `null` for querying tax entities and fall back to using `0` as the tax rate instead.